// src/api/model/pollQuestionsUpsertRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PollQuestionsUpsertRequestQuestionsItem } from './pollQuestionsUpsertRequestQuestionsItem';

export interface PollQuestionsUpsertRequest {
  questions: PollQuestionsUpsertRequestQuestionsItem[];
}