// src/api/model/pollResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PollResponseStatus } from './pollResponseStatus';
import type { PollResponseQuestionsItem } from './pollResponseQuestionsItem';

export interface PollResponse {
  id: number;
  createdDate: string;
  /**
   * @minLength 0
   * @maxLength 255
   */
  name: string;
  description?: string;
  status: PollResponseStatus;
  questions: PollResponseQuestionsItem[];
  mySubmissionsCount: number;
}